Reorganized database utilities & added flask code
Flask startup code is located in startup.py
