--- CREATE_DATABASE.sql
@@ -1,7 +1,0 @@
-CREATE DATABASE TTOps;
-
-SHOW DATABASES;
-
-CREATE USER 'TTAdmin' IDENTIFIED BY 'ibuprofen-fiddle-wound3';
-GRANT ALL PRIVILEGES ON TTOps.* TO 'TTADMIN';
-FLUSH PRIVILEGES;
