--- CREATE_VW_INVENTORY_ARCHIVE_SUMMARY.sql
@@ -1,3 +1,0 @@
-CREATE VIEW vw_InventoryArchiveSummary AS
-SELECT ArchiveID, SKU, ItemName, ItemPrice, ItemQuantity, Supplier, FinalUpdateAt, ArchivedDate
-FROM InventoryArchive;
