--- CREATE_VW_USER_ROLES.sql
@@ -1,3 +1,0 @@
-CREATE VIEW vw_UserRoles AS
-SELECT UserID, FirstName, LastName, Email, Role, CreatedAt, LastLogin
-FROM Users;
